requirements.txt
toml
gitpython

# development
mypy